Disable AFF4 volume when extension is not .aff4. (#43)
When outputting to raw or ELF it is expected that the output be a flat
file and not an AFF4 volume. Unless the filename ends with .aff4, this
will now be a flat file.

When writing to a flat file we can not append any other streams. This
is obviously less useful.